+++ packages/domain/src/garantiesFinancières/dépôt/déposerGarantiesFinancières.command.ts
@@ -1,58 +1,60 @@
 import { Message, MessageHandler, mediator } from 'mediateur';
-import { IdentifiantProjetValueType } from '../projet/projet.valueType';
+import { IdentifiantProjetValueType } from '../../projet/projet.valueType';
 import { LoadAggregate, Publish } from '@potentiel/core-domain';
-import { verifyGarantiesFinancièresTypeForCommand } from '../projet/garantiesFinancières/verifyGarantiesFinancièresTypeForCommand';
-import { verifyGarantiesFinancièresAttestationForCommand } from '../projet/garantiesFinancières/verifyGarantiesFinancièresAttestationForCommand';
+import { verifyGarantiesFinancièresTypeForCommand } from '../../projet/garantiesFinancières/verifyGarantiesFinancièresTypeForCommand';
+import { verifyGarantiesFinancièresAttestationForCommand } from '../../projet/garantiesFinancières/verifyGarantiesFinancièresAttestationForCommand';
 import {
   AttestationConstitution,
   TypeEtDateÉchéance,
   estTypeAvecDateÉchéance,
-} from './garantiesFinancières.valueType';
-import { Utilisateur } from '../domain.valueType';
+} from '../garantiesFinancières.valueType';
+import { DateTimeValueType, Utilisateur } from '../../domain.valueType';
 import {
   createDépôtGarantiesFinancièresAggregateId,
   loadDépôtGarantiesFinancièresAggregateFactory,
-} from './garantiesFinancières.aggregate';
-import { TéléverserFichierPort } from '../common.ports';
-import { isNone } from '@potentiel/monads';
-import { DépôtGarantiesFinancièresModifiéV1 } from './garantiesFinancières.event';
-import { DépôtGarantiesFinancièresNonTrouvéPourModificationErreur } from './garantiesFinancières.error';
+} from '../garantiesFinancières.aggregate';
+import { GarantiesFinancièresDéposéesV1 } from '../garantiesFinancières.event';
+import { TéléverserFichierPort } from '../../common.ports';
+import { isSome } from '@potentiel/monads';
+import { DépôtGarantiesFinancièresDéjàExistantErreur } from '../garantiesFinancières.error';
 
-export type ModifierDépôtGarantiesFinancièresCommand = Message<
-  'MODIFIER_DÉPÔT_GARANTIES_FINANCIÈRES',
+export type DéposerGarantiesFinancièresCommand = Message<
+  'DÉPOSER_GARANTIES_FINANCIÈRES',
   {
     identifiantProjet: IdentifiantProjetValueType;
     attestationConstitution: AttestationConstitution;
     utilisateur: Utilisateur;
+    dateDépôt: DateTimeValueType;
   } & TypeEtDateÉchéance
 >;
 
-export type ModifierDépôtGarantiesFinancièresDependencies = {
+export type DéposerGarantiesFinancièresDependencies = {
   publish: Publish;
   loadAggregate: LoadAggregate;
   téléverserFichier: TéléverserFichierPort;
 };
 
-export const registerModifierDépôtGarantiesFinancièresCommand = ({
+export const registerDéposerGarantiesFinancièresCommand = ({
   publish,
   loadAggregate,
   téléverserFichier,
-}: ModifierDépôtGarantiesFinancièresDependencies) => {
+}: DéposerGarantiesFinancièresDependencies) => {
   const loadDépôtGarantiesFinancières = loadDépôtGarantiesFinancièresAggregateFactory({
     loadAggregate,
   });
 
-  const handler: MessageHandler<ModifierDépôtGarantiesFinancièresCommand> = async ({
+  const handler: MessageHandler<DéposerGarantiesFinancièresCommand> = async ({
     identifiantProjet,
     typeGarantiesFinancières,
     dateÉchéance,
     attestationConstitution,
+    dateDépôt,
     utilisateur,
   }) => {
     const agrégatDépôtGarantiesFinancières = await loadDépôtGarantiesFinancières(identifiantProjet);
 
-    if (isNone(agrégatDépôtGarantiesFinancières) || !agrégatDépôtGarantiesFinancières.dépôt) {
-      throw new DépôtGarantiesFinancièresNonTrouvéPourModificationErreur();
+    if (isSome(agrégatDépôtGarantiesFinancières) && agrégatDépôtGarantiesFinancières.dépôt) {
+      throw new DépôtGarantiesFinancièresDéjàExistantErreur();
     }
 
     verifyGarantiesFinancièresTypeForCommand(typeGarantiesFinancières, dateÉchéance, utilisateur);
@@ -66,8 +68,8 @@ export const registerModifierDépôtGarantiesFinancièresCommand = ({
       type: 'depot-attestation-constitution-garanties-financieres',
     });
 
-    const event: DépôtGarantiesFinancièresModifiéV1 = {
-      type: 'DépôtGarantiesFinancièresModifié-v1',
+    const event: GarantiesFinancièresDéposéesV1 = {
+      type: 'GarantiesFinancièresDéposées-v1',
       payload: {
         identifiantProjet: identifiantProjet.formatter(),
         ...(estTypeAvecDateÉchéance(typeGarantiesFinancières)
@@ -80,11 +82,12 @@ export const registerModifierDépôtGarantiesFinancièresCommand = ({
           format: attestationConstitution.format,
           date: attestationConstitution.date.formatter(),
         },
+        dateDépôt: dateDépôt.formatter(),
       },
     };
 
     await publish(createDépôtGarantiesFinancièresAggregateId(identifiantProjet), event);
   };
 
-  mediator.register('MODIFIER_DÉPÔT_GARANTIES_FINANCIÈRES', handler);
+  mediator.register('DÉPOSER_GARANTIES_FINANCIÈRES', handler);
 };
